chore: move shared flag logic to its own file, helpers.go feels too generic for this code
